LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-window-open-self-expected.txt


FAIL window.open() fires navigate event when targeting self assert_equals: expected (object) null but got (string) ""
